manit_bus/requirements.txt
Django>=4.0,<5.0
djangorestframework
djangorestframework-simplejwt
gunicorn
whitenoise
django-cors-headers
